+++ R/seltailor.R
@@ -5,7 +5,7 @@
 #' @usage seltailor(X = NULL, wl = 512, flim = c(0,22), wn = "hanning", mar = 0.5,
 #'  osci = TRUE, pal = reverse.gray.colors.2, ovlp = 70, auto.next = FALSE, pause = 1,
 #'   comments = TRUE, path = NULL, frange = FALSE, fast.spec = FALSE, ext.window = TRUE,
-#'   width = 15, height = 5, ...)
+#'   width = 15, height = 5, index = NULL, ...)
 #' @param X data frame with the following columns: 1) "sound.files": name of the .wav 
 #' files, 2) "selec": number of the selections, 3) "start": start time of selections, 4) "end": 
 #' end time of selections. The ouptut of \code{\link{manualoc}} or \code{\link{autodetec}} can 
@@ -48,6 +48,9 @@
 #' if \code{ext.window = FALSE}. Default is 15.
 #' @param height Numeric of length 1 controling the height of the external graphic window.
 #' Ignored if \code{ext.window = FALSE}. Default is 5.
+#' @param index Numeric vector indicating which selections (rows) of 'X' should be tailored. 
+#'  Default is \code{NULL}. Ignored when the process is resumed. This can be useful when combined
+#'  with \code{\link{filtersels}}) output (see 'index' argument in \code{\link{filtersels}}).
 #' @param ... Additional arguments to be passed to the internal spectrogram creating function for customizing graphical output. The function is a modified version of \code{\link[seewave]{spectro}}, so it takes the same arguments. 
 #' @return .csv file saved in the working directory with start and end time of 
 #'   selections.
@@ -104,7 +107,7 @@
 seltailor <- function(X = NULL, wl = 512, flim = c(0,22), wn = "hanning", mar = 0.5,
             osci = TRUE, pal = reverse.gray.colors.2, ovlp = 70, auto.next = FALSE,
             pause = 1, comments = TRUE, path = NULL, frange = FALSE, fast.spec = FALSE,
-            ext.window = TRUE, width = 15, height = 5, ...)
+            ext.window = TRUE, width = 15, height = 5, index = NULL, ...)
 {
  
   #X must be provided
@@ -138,15 +141,18 @@ seltailor <- function(X = NULL, wl = 512, flim = c(0,22), wn = "hanning", mar = 
   if(!file.exists(file.path(getwd(), "seltailor_output.csv")))
   {X$tailored <- ""
   X$tailored <- as.character(X$tailored)
-    write.csv(X, "seltailor_output.csv", row.names =  FALSE)  
+  if(!is.null(index))   X$tailored[!1:nrow(X) %in% index] <- "y"
+  write.csv(X, "seltailor_output.csv", row.names =  FALSE)  
     } else {X <- read.csv("seltailor_output.csv", stringsAsFactors = FALSE)  
   if(any(is.na(X$tailored))) X$tailored[is.na(X$tailored)] <-""
   if(all(any(!is.na(X$tailored)),X$tailored[nrow(X)] == "y")) stop("all selections have been analyzed") 
     }
   
-  dn <- 0
-  if(any(!is.na(X$tailored))) if(length(which(X$tailored == "y")) > 0) dn = max(which(X$tailored == "y"))
-   
+  
+  dn <- 1:nrow(X)
+  if(any(!is.na(X$tailored))) if(length(which(X$tailored == "y")) > 0) 
+    dn <- which(X$tailored != "y")  
+
   #set external window function
   if(any(Sys.info()[1] == c("Linux", "Windows"))) extwin <- grDevices::X11 else extwin <- grDevices::quartz
   
@@ -154,7 +160,7 @@ seltailor <- function(X = NULL, wl = 512, flim = c(0,22), wn = "hanning", mar = 
   if(ext.window)  extwin(width = width, height = height)
   
   #this first loop runs over selections
-  for(j in (dn + 1):nrow(X))
+  for(j in dn)
     {
     rec <- tuneR::readWave(as.character(X$sound.files[j]), header = TRUE)
     main <- paste(X$sound.files[j], X$selec[j], sep = "-") 
@@ -222,7 +228,7 @@ seltailor <- function(X = NULL, wl = 512, flim = c(0,22), wn = "hanning", mar = 
         X$tailored[j] <- "y"
         write.csv(X, "seltailor_output.csv", row.names =  FALSE)  
         
-        if(X$tailored[nrow(X)] == "y") {
+        if(nrow(X[X$tailored== "y",]) == nrow(X)) {
           dev.off()
           stop("all selections have been analyzed")
           } 
@@ -293,7 +299,7 @@ seltailor <- function(X = NULL, wl = 512, flim = c(0,22), wn = "hanning", mar = 
           if(min(xy3$y) < 0) X$low.f[j] <- 0  
           X$high.f[j] <- max(xy3$y)
         }
-     if(X$tailored[nrow(X)] == "y") { dev.off()
+     if(nrow(X[X$tailored== "y",]) == nrow(X)) { dev.off()
        stop("all selections have been analyzed") 
        } else
       {Sys.sleep(pause)